+++ src/main/kotlin/io/github/inductiveautomation/kindling/MainPanel.kt
@@ -336,8 +336,8 @@ class MainPanel : JPanel(MigLayout("ins 6, fill")) {
 
                 jFrame(
                     title = "Kindling",
-                    width = 1680,
-                    height = 1050,
+                    width = 1280,
+                    height = 800,
                     embedContentIntoTitleBar = true,
                 ) {
                     defaultCloseOperation = JFrame.EXIT_ON_CLOSE
