+++ src/main/kotlin/io/github/inductiveautomation/kindling/utils/NoSelectionModel.kt
@@ -1,0 +1,24 @@
+package io.github.inductiveautomation.kindling.utils
+
+import javax.swing.DefaultListSelectionModel
+import javax.swing.ListSelectionModel
+
+/**
+ * A simple [ListSelectionModel] implementation that never allows selecting any elements.
+ */
+class NoSelectionModel : DefaultListSelectionModel() {
+    override fun setSelectionInterval(index0: Int, index1: Int) = Unit
+    override fun addSelectionInterval(index0: Int, index1: Int) = Unit
+    override fun removeSelectionInterval(index0: Int, index1: Int) = Unit
+    override fun getMinSelectionIndex(): Int = -1
+    override fun getMaxSelectionIndex(): Int = -1
+    override fun isSelectedIndex(index: Int): Boolean = false
+    override fun getAnchorSelectionIndex(): Int = -1
+    override fun setAnchorSelectionIndex(index: Int) = Unit
+    override fun getLeadSelectionIndex(): Int = -1
+    override fun setLeadSelectionIndex(index: Int) = Unit
+    override fun clearSelection() = Unit
+    override fun isSelectionEmpty(): Boolean = true
+    override fun insertIndexInterval(index: Int, length: Int, before: Boolean) = Unit
+    override fun removeIndexInterval(index0: Int, index1: Int) = Unit
+}
